move publish interfaces out of sqldbproj.d.ts
